Added option forceLinearModel and adapted convexProblem for the specific direction cases of prodIndex
prodIndex works with specific directions for which even the convex, multiplicative case, that in the general
form is non-linear, can be rewritten as a linear problem by just variable substitution invLambda = 1/lambda.

This commit add the option `forceLinearModel` to `problem` and `convexProblem` so that prodIndex can use it,
and readjust the convex multiplicative case for the specific direction cases that can arise in prodIndex.